passing down players list from state, randomising and rendering pairs, only for 4
