+++ src/containers/Round.js
@@ -1,0 +1,15 @@
+// ==========Import React redux's connect function
+import { connect } from "react-redux";
+
+//===========Import in the component to wrap
+import Round from "../components/show/Round";
+
+// mapStateToProps is passed in the current state returns object, which gets passed in as props to the connected component
+const mapStateToProps = state => {
+    return {
+        players: state.get("players"),
+    };
+};
+
+// connect up mapStateToProps with the Round component
+export default connect(mapStateToProps)(Round);
