+++ src/data/initial.js
@@ -1,7 +1,7 @@
 import { Map, List } from "immutable";
 
 const initial = Map({
-	players: List([]),
+	players: List(["p1", "p2", "p3", "p4"]),
 });
 
 export default initial;
